# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laroche API Laravel 12",
        "description": "Documentação completa da API Laravel com Sanctum, Swagger e uma arquitetura limpa (SOLID).",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cadastro de um novo administrador",
                "operationId": "d6f5d00c40f9de2e9272088aebe648c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Super Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conta criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login de um administrador",
                "operationId": "0f5784f47a3fcac35f86ec4e82d967d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conectado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Adicionar um comentário em um perfil (somente administradores, apenas um comentário por perfil)",
                "operationId": "943ef5a9a3a98223c01a1267e636d7e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "profile_id",
                                    "content"
                                ],
                                "properties": {
                                    "profile_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Ótimo perfil!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comentário adicionado com sucesso"
                    },
                    "403": {
                        "description": "Ação não permitida: este perfil já possui um comentário."
                    },
                    "422": {
                        "description": "Alguns campos estão incorretos ou ausentes"
                    },
                    "401": {
                        "description": "Usuário não autenticado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Recuperar a lista de perfis ativos (público, status visível apenas para administradores autenticados)",
                "operationId": "92d33432e4b6b56fbf6887fede4b2a66",
                "responses": {
                    "200": {
                        "description": "Liste des profils actifs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nom": {
                                                "type": "string",
                                                "example": "Laroche"
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Henri"
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "http://localhost:8000/storage/profiles/profile.jpg"
                                            },
                                            "admin_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "status": {
                                                "description": "Visível apenas para administradores autenticados",
                                                "type": "string",
                                                "example": "actif"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado (opcional, somente se necessário)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Créer un profil (admin uniquement)",
                "operationId": "c0c170a3a5458f63601fcd9ec2fe896a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "first_name",
                                    "image",
                                    "status"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Laroche"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Henri"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "inactif",
                                            "en attente",
                                            "actif"
                                        ],
                                        "example": "actif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil créé avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Non autorisé (rôle insuffisant)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Token Sanctum (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Comments",
            "description": "Comments"
        },
        {
            "name": "Profiles",
            "description": "Profiles"
        }
    ]
}